requirements.txt
absl-py==1.0.0
adjustText==0.7.3

anaconda-navigator==2.0.3
appdirs==1.4.4
argh==0.26.2
astunparse==1.6.3
atomicwrites==1.4.0

backports.weakref==1.0.post1

bioinfokit==2.0.8
bkcharts==0.2
black==19.10b0

boto==2.49.0
Bottleneck==1.3.2
brotlipy==0.7.0
cachetools==4.2.4
certifi==2020.12.5

clyent==1.2.2
colour==0.1.5
comtypes==1.1.9
conda==4.10.3
conda-build==3.21.4

conda-verify==3.4.2
contextlib2==0.6.0.post1
cycler==0.10.0
Cython==0.29.28
cytoolz==0.11.0

deepface==0.0.70

docopt==0.6.2
ensemble==0.0.dev1
entrypoints==0.3
et-xmlfile==1.0.1
factor-analyzer==0.4.0
fastcache==1.1.0

flatbuffers==2.0
future==0.18.2
gast==0.3.3
gdown==4.2.0
gensim==4.2.0

google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
h5py==2.10.0
HeapDict==1.0.1

imbalanced-learn==0.8.1
imblearn==0.0
importlib-metadata==4.8.2

jdcal==1.4.1

json5==0.9.5
jupyter==1.0.0

jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1

jupyterthemes==0.20.0
keras==2.8.0
Keras-Preprocessing==1.1.2

lesscpy==0.15.0
libclang==12.0.0
llvmlite==0.36.0
locket==0.2.1
Markdown==3.3.6
MarkupSafe==1.1.1
matplotlib-venn==0.11.7
mccabe==0.6.1
menuinst==1.4.16
mistune==0.8.4
mkl-fft==1.3.0

mkl-service==2.3.0
mlxtend==0.19.0

mpmath==1.2.1
mtcnn==0.1.1
multipledispatch==0.6.0
mypy-extensions==0.4.3
mysql-connector-python==8.0.29
navigator-updater==0.2.1

numpy==1.23.0
oauthlib==3.1.1
olefile==0.46
opencv-python==4.5.5.62
opt-einsum==3.3.0

pandas-datareader==0.10.0

parso==0.7.0

pathspec==0.7.0
patsy==0.5.1
pep8==1.7.1

pkginfo==1.7.0
ply==3.11
pmdarima==1.8.5

protobuf==3.19.1
Python==3.10.5

pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycurl==7.43.0.6

pyodbc===4.0.0-unsupported

PyPrind==2.11.3
pyreadline==2.1

pysut==1.1
pytest==6.2.3

pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==20.0.0
QDarkStyle==2.8.1

QtPy==1.9.0

requests-oauthlib==1.3.0
retina-face==0.0.7
rsa==4.8

scikit-image==0.18.1
scikit-learn==1.1.0
scipy==1.4.1

simplegeneric==0.8.1
sip==4.19.13
sklearn==0.0
smart-open==6.0.0

stats-arrays==0.6.5
statsmodels==0.12.2
tables==3.6.1
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.8.0
tensorflow-estimator==2.2.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.9.4
textblob==0.17.1
textwrap3==0.9.2
tf-estimator-nightly==2.8.0.dev2021122109

unicodecsv==0.14.1

webencodings==0.5.1
widgetsnbextension==3.5.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.12.1
xgboost==1.5.1

xlwings==0.23.0
xlwt==1.3.0
xmltodict==0.12.0
yellowbrick==1.4
zict==2.0.0
zope.event==4.5.0
